Fixed Timespan and Location bug.
Timespans now use Longs. Which means it is not limited to 24 days or
something. (TESTED)

The quick fix for a Location bug works now.

Code used to test. http://pastebin.com/wEL9fzdf

+++ src/main/java/ch/njol/skript/classes/data/SkriptClasses.java
@@ -419,7 +419,7 @@ public class SkriptClasses {
 					@Nullable
 					public Timespan deserialize(final String s) {
 						try {
-							return new Timespan(Integer.parseInt(s));
+							return new Timespan(Long.parseLong(s));
 						} catch (final NumberFormatException e) {
 							return null;
 						}
